moving relevant flles into wf_parser
