+++ wf_field/src/lib.rs
@@ -14,7 +14,6 @@ pub mod definitions;
 mod field;
 mod field_definition;
 mod field_definition_parser;
-mod header;
 mod request;
 mod types;
 
@@ -24,8 +23,8 @@ mod test_field_definition;
 pub use {
     field::Field,
     field_definition::FieldDefinition,
-    field_definition_parser::{FieldDefinitionParser, Parser},
-    header::MessageHeaderOrder,
+    field_definition_parser::{FieldDefinitionParser, FieldDefinitionParserBase},
+    request::create_request_fields,
     types::MessageType,
 };
 
@@ -53,4 +52,26 @@ pub trait MessageHeader {
     fn message_code(&self) -> &Self::Target;
     fn reference_indicator(&self) -> &Self::Target;
     fn referenced_message(&self) -> &Self::Target;
+}
+
+#[repr(usize)]
+#[derive(Copy, Clone)]
+pub enum MessageHeaderOrder {
+    Prefix = 0,
+    Version = 1,
+    EncryptionIndicator = 2,
+    DuressIndicator = 3,
+    MessageCode = 4,
+    ReferenceIndicator = 5,
+    ReferencedMessage = 6,
+}
+
+impl<'a> MessageHeaderOrder {
+    pub fn as_usize(&self) -> usize {
+        *self as usize
+    }
+
+    pub fn get<'b>(&'a self, fields: &'b [Field]) -> &'b Field {
+        &fields[self.as_usize()]
+    }
 }
